feat(decorators,swagger-generator)!: move preset and flatten generate API
The metadata preset now lives where it's most useful — next to the
decorators it decodes — and is published as the `./preset` subpath
export of `@routup/decorators`. Any trapi-based tool (the `trapi` CLI,
custom generators) can consume it directly via
`--preset @routup/decorators/preset` without depending on
`@routup/swagger-generator`.

`@routup/decorators` gains `@trapi/metadata` as an *optional* peer
dependency (peerDependenciesMeta.optional = true) — runtime-only users
who never touch the preset don't need it installed.

`@routup/swagger-generator` is now a thin wrapper:
- `generate()` takes `SwaggerGenerateOptions` directly
  (`{ version, metadata, data }`), the same shape as
  `@trapi/swagger`'s `generateSwagger()` — no more `GeneratorContext`
  wrapper, no more `options.metadata` / `options.<data fields>` mix.
- `buildPreset()` is re-exported from `@routup/decorators/preset` so
  it's still a single import for typical consumers, and the rest of
  `@trapi/swagger`'s public API (`Version`, `SpecV2`/`V3`,
  `saveSwagger`, types) is re-exported via `export * from
  '@trapi/swagger'`.
- The wrapper applies the same defaults as before (entryPoint =
  `<cwd>/src/**/*.ts`, ignore node_modules, preset = buildPreset(),
  version = V3_2, default data name/description/consumes/produces).

BREAKING CHANGE: `@routup/swagger-generator` no longer exports
`GeneratorContext`, `GeneratorOptionsInput`, or `GeneratorOutput`, and
`generate()`'s call shape changed. Migrate from
`generate({ version, options: { metadata, ...data } })` to
`generate({ version, metadata, data })`. The internal preset name
moved from `@routup/swagger-generator` to `@routup/decorators` (only
relevant if you cache by preset name).
