+++ packages/decorators/package.json
@@ -8,6 +8,10 @@
         ".": {
             "types": "./dist/index.d.mts",
             "import": "./dist/index.mjs"
+        },
+        "./preset": {
+            "types": "./dist/preset.d.mts",
+            "import": "./dist/preset.mjs"
         }
     },
     "types": "./dist/index.d.mts",
@@ -51,7 +55,13 @@
     },
     "homepage": "https://github.com/routup/plugins#readme",
     "peerDependencies": {
+        "@trapi/metadata": "^2.0.0-beta.2",
         "routup": "^5.0.0-beta.6"
+    },
+    "peerDependenciesMeta": {
+        "@trapi/metadata": {
+            "optional": true
+        }
     },
     "dependencies": {
         "@routup/body": "^2.4.3",
@@ -59,6 +69,7 @@
         "@routup/query": "^2.4.3"
     },
     "devDependencies": {
+        "@trapi/metadata": "^2.0.0-beta.2",
         "routup": "^5.0.0-beta.6"
     },
     "publishConfig": {
