+++ packages/swagger-generator/src/generate.ts
@@ -1,23 +1,32 @@
 import path from 'node:path';
 import process from 'node:process';
+import { buildPreset } from '@routup/decorators/preset';
 import { isMetadata } from '@trapi/metadata';
-import { Version, generateSwagger } from '@trapi/swagger';
+import {
+    Version,
+    generateSwagger,
+} from '@trapi/swagger';
+import type {
+    OutputForVersion,
+    SwaggerGenerateData,
+    SwaggerGenerateOptions,
+} from '@trapi/swagger';
 import { createMerger } from 'smob';
-import type { GeneratorContext, GeneratorOptionsInput, GeneratorOutput } from './type';
-import { buildPreset } from '../preset';
 
-const DEFAULT_DATA = {
+const DEFAULT_DATA: SwaggerGenerateData = {
     name: 'API Documentation',
     description: 'Explore the REST Endpoints of the API.',
     consumes: ['application/json'],
     produces: ['application/json'],
 };
 
-export async function generate<V extends `${Version}`>(
-    context: GeneratorContext<V>,
-): Promise<GeneratorOutput<V>> {
+export type GenerateOptions<V extends `${Version}` = `${Version}`> = Partial<Omit<SwaggerGenerateOptions, 'version'>> & { version?: V };
+
+export async function generate<V extends `${Version}` = typeof Version.V3_2>(
+    options: GenerateOptions<V> = {},
+): Promise<OutputForVersion<V>> {
     const merge = createMerger({ array: true, arrayDistinct: true });
-    const options = merge({}, DEFAULT_DATA, context.options || {}) as GeneratorOptionsInput;
+    const data = merge({}, DEFAULT_DATA, options.data || {}) as SwaggerGenerateData;
 
     let { metadata } = options;
     if (!metadata) {
@@ -28,9 +37,7 @@ export async function generate<V extends `${Version}`>(
                 { pattern: '**/*.ts', cwd: path.join(process.cwd(), 'src') },
             ],
         };
-    }
-
-    if (!isMetadata(metadata)) {
+    } else if (!isMetadata(metadata)) {
         if (!metadata.entryPoint) {
             metadata.entryPoint = [
                 { pattern: '**/*.ts', cwd: path.join(process.cwd(), 'src') },
@@ -39,17 +46,11 @@ export async function generate<V extends `${Version}`>(
         if (!metadata.preset) {
             metadata.preset = buildPreset();
         }
-        if (context.tsconfig && !metadata.tsconfig) {
-            metadata.tsconfig = context.tsconfig;
-        }
     }
 
-    // Strip metadata from `data` so it isn't fed back as a spec extension.
-    const { metadata: _omit, ...data } = options;
-
-    return await generateSwagger({
-        version: context.version || Version.V3_2,
+    return generateSwagger({
+        version: (options.version || Version.V3_2) as V,
         metadata,
         data,
-    }) as GeneratorOutput<V>;
+    });
 }
