--- packages/swagger-generator/src/generator/constants.ts
@@ -1,5 +1,0 @@
-import { Version } from '@trapi/swagger';
-
-export {
-    Version,
-};
