--- packages/swagger-generator/src/generator/index.ts
@@ -1,3 +1,0 @@
-export * from './constants';
-export * from './module';
-export * from './type';
