+++ packages/swagger-generator/src/index.ts
@@ -1,2 +1,10 @@
-export * from './generator';
-export * from './preset';
+export type {
+    Metadata,
+    MetadataGenerateOptions,
+    Preset,
+    TsCompilerOptions,
+    TsConfig,
+} from '@trapi/metadata';
+export * from '@trapi/swagger';
+export { buildPreset } from '@routup/decorators/preset';
+export * from './generate';
